feat!: decode Evidence from Read
Change Evidence::decode to accept a generic type that is Read. This
allows for more flexibility in how this API is used, removing the need
for additional conversions, and in some cases (e.g. when reading from
a file), removing the need to created an intermediate buffer entirely.

BREAKING_CHANGE: Signature of Evidence::decode is changed.

Signed-off-by: setrofim <setrofim@gmail.com>